rename pyshiny -> shiny
